Changes requested by Cornelia to accommodate distance education students
Changed "Are you a faculty member or graduate student?" to “Status” and added Undergraduate and Distance student as options.  Once the form label is changed to just Status, I think it makes sense to include the undergraduate status (even though it doesn’t do anything) just to keep undergraduates from getting confused.

Changed “Pickup library” to “Pickup location”.

"Delivery to address" is added to Pickup location when Distance student is selected in the Status field.

Added a note that displays when Delivery to address is selected.